refactor(data-model): move source into src/, split benches into bench/
Reorganize the `data-model` package directory layout to match the
`utils` package convention:

- All non-test/non-bench source `.ts` files move into `src/`.
- `deno.json` stays at the package root; its export RHS paths are
  repointed `./Foo.ts` -> `./src/Foo.ts`.
- Unit tests stay in `test/`; their relative imports of source are
  repointed `../foo.ts` -> `../src/foo.ts` (static and dynamic).
- Benchmarks move into a new `bench/` directory:
  `fabric-value.bench.ts` relocates as-is, and `hashing-bench.ts`
  becomes `bench/hashing.bench.ts` (relocation plus the `.bench.ts`
  naming convention in one move). A `bench` task is added to
  `deno.json`.

Cross-package importers are unaffected: they use the bare
`@commonfabric/data-model/<name>` export alias, so the directory
move is transparent once the `deno.json` export paths are repointed.

Conservation of content: source files move byte-identical (git
detects them as 100%-similarity renames); only test and bench files
change, and only to repoint import paths.
