+++ packages/data-model/bench/hashing.bench.ts
@@ -1,11 +1,11 @@
 /**
  * Hashing Performance Benchmarks
  *
- * Run with: deno bench --allow-read --allow-write --allow-net --allow-ffi --allow-env --no-check test/hashing-bench.ts
+ * Run with: deno bench --allow-read --allow-write --allow-net --allow-ffi --allow-env --no-check bench/hashing.bench.ts
  */
 
-import { hashOf } from "../value-hash.ts";
-import { deepFreeze } from "../deep-freeze.ts";
+import { hashOf } from "../src/value-hash.ts";
+import { deepFreeze } from "../src/deep-freeze.ts";
 
 // ---------------------------------------------------------------------------
 // Pre-generated test data
