+++ packages/data-model/src/FabricBytes.ts
