+++ packages/data-model/src/deep-freeze.ts
