+++ packages/data-model/src/fabric-type-tags.ts
