+++ packages/data-model/src/fabric-value-legacy.ts
