+++ packages/data-model/src/fabric-value.ts
