+++ packages/data-model/src/interface.ts
