+++ packages/data-model/src/json-encoding.ts
