+++ packages/data-model/test/clone-for-mutation.test.ts
@@ -5,11 +5,11 @@ import {
   CloneForMutationError,
   resetDataModelConfig,
   setDataModelConfig,
-} from "../fabric-value.ts";
-import type { FabricValue } from "../fabric-value.ts";
-import { deepFreeze, isDeepFrozen } from "../deep-freeze.ts";
-import { FabricEpochNsec } from "../fabric-epoch.ts";
-import { FabricError } from "../fabric-native-instances.ts";
+} from "../src/fabric-value.ts";
+import type { FabricValue } from "../src/fabric-value.ts";
+import { deepFreeze, isDeepFrozen } from "../src/deep-freeze.ts";
+import { FabricEpochNsec } from "../src/fabric-epoch.ts";
+import { FabricError } from "../src/fabric-native-instances.ts";
 
 // ============================================================================
 // `cloneForMutation` tests
