+++ packages/data-model/test/fabric-regexp.test.ts
@@ -5,25 +5,25 @@ import {
   FabricInstance,
   type FabricValue,
   RECONSTRUCT,
-} from "../interface.ts";
-import { BaseReconstructionContext } from "../BaseReconstructionContext.ts";
+} from "../src/interface.ts";
+import { BaseReconstructionContext } from "../src/BaseReconstructionContext.ts";
 import {
   FabricNativeWrapper,
   FabricRegExp,
   isConvertibleNativeInstance,
-} from "../fabric-native-instances.ts";
+} from "../src/fabric-native-instances.ts";
 import {
   isFabricCompatible,
   resetDataModelConfig,
   setDataModelConfig,
   shallowFabricFromNativeValue,
-} from "../fabric-value.ts";
+} from "../src/fabric-value.ts";
 import {
   NATIVE_TAGS,
   tagFromNativeClass,
   tagFromNativeValue,
-} from "../native-type-tags.ts";
-import { hashOf } from "../value-hash.ts";
+} from "../src/native-type-tags.ts";
+import { hashOf } from "../src/value-hash.ts";
 
 /** Dummy reconstruction context for tests. */
 class DummyReconstructionContext extends BaseReconstructionContext {
